consolidate execution and verification
