--- contracts/scripts/testVerifyBoldUpgrade.ts
@@ -1,129 +1,0 @@
-import { ethers } from 'ethers'
-import { DeployedContracts, getConfig, getJsonFile } from './common'
-import dotenv from 'dotenv'
-import {
-  EdgeChallengeManager__factory,
-  RollupUserLogic__factory,
-} from '../build/types'
-import { getAddress } from 'ethers/lib/utils'
-
-dotenv.config()
-
-async function main() {
-  const l1RpcVal = process.env.L1_RPC_URL
-  if (!l1RpcVal) {
-    throw new Error('L1_RPC_URL env variable not set')
-  }
-  const l1Rpc = new ethers.providers.JsonRpcProvider(l1RpcVal)
-
-  const configLocation = process.env.CONFIG_LOCATION
-  if (!configLocation) {
-    throw new Error('CONFIG_LOCATION env variable not set')
-  }
-  const config = await getConfig(configLocation, l1Rpc)
-
-  const deployedContractsLocation = process.env.DEPLOYED_CONTRACTS_LOCATION
-  if (!deployedContractsLocation) {
-    throw new Error('DEPLOYED_CONTRACTS_LOCATION env variable not set')
-  }
-
-  // load the deployed contracts
-  const existingDeployedContracts = getJsonFile(
-    deployedContractsLocation
-  ) as DeployedContracts
-
-  if (!existingDeployedContracts.newEdgeChallengeManager) {
-    throw new Error(
-      "newEdgeChallengeManager doesn't exist in deployed contracts"
-    )
-  }
-
-  const edgeChallengeManager = EdgeChallengeManager__factory.connect(
-    existingDeployedContracts.newEdgeChallengeManager,
-    l1Rpc
-  )
-  if (getAddress(await edgeChallengeManager.stakeToken()) != getAddress(config.settings.stakeToken)) {
-    throw new Error('Stake token address does not match')
-  }
-
-  for (let i = 0; i < config.settings.miniStakeAmounts.length; i++) {
-    if (
-      !(await edgeChallengeManager.stakeAmounts(i)).eq(
-        config.settings.miniStakeAmounts[i]
-      )
-    ) {
-      throw new Error('Mini stake amount does not match')
-    }
-  }
-
-  if (
-    !(await edgeChallengeManager.challengePeriodBlocks()).eq(
-      config.settings.challengePeriodBlocks
-    )
-  ) {
-    throw new Error('Challenge period blocks does not match')
-  }
-
-  if (
-    !(await edgeChallengeManager.LAYERZERO_BLOCKEDGE_HEIGHT()).eq(
-      config.settings.blockLeafSize
-    )
-  ) {
-    throw new Error('Block leaf size does not match')
-  }
-
-  if (
-    !(await edgeChallengeManager.LAYERZERO_BIGSTEPEDGE_HEIGHT()).eq(
-      config.settings.bigStepLeafSize
-    )
-  ) {
-    throw new Error('Big step leaf size does not match')
-  }
-
-  if (
-    !(await edgeChallengeManager.LAYERZERO_SMALLSTEPEDGE_HEIGHT()).eq(
-      config.settings.smallStepLeafSize
-    )
-  ) {
-    throw new Error('Small step leaf size does not match')
-  }
-
-  if (
-    (await edgeChallengeManager.NUM_BIGSTEP_LEVEL()) !==
-    config.settings.numBigStepLevel
-  ) {
-    throw new Error('Number of big step level does not match')
-  }
-
-  const assertionChain = RollupUserLogic__factory.connect(
-    await edgeChallengeManager.assertionChain(),
-    l1Rpc
-  )
-
-  if (getAddress(await assertionChain.stakeToken()) != getAddress(config.settings.stakeToken)) {
-    throw new Error('Stake token address does not match')
-  }
-
-  if (
-    !(await assertionChain.confirmPeriodBlocks()).eq(
-      config.settings.confirmPeriodBlocks
-    )
-  ) {
-    throw new Error('Confirm period blocks does not match')
-  }
-
-  if (!(await assertionChain.baseStake()).eq(config.settings.stakeAmt)) {
-    throw new Error('Base stake does not match')
-  }
-
-  if (config.settings.anyTrustFastConfirmer.length != 0) {
-    if (
-      getAddress(await assertionChain.anyTrustFastConfirmer()) !=
-      getAddress(config.settings.anyTrustFastConfirmer)
-    ) {
-      throw new Error('Any trust fast confirmer does not match')
-    }
-  }
-}
-
-main().then(() => console.log('Done.'))
